fix(KONFLUX-6218): align repository ids to cpe mapping
This update changes the rpm repository ids to match those found in Red Hat's
repository-to-cpe.json mapping file[0], used by third-party scanners.

In order for scanners like clair to understand what CPE[1] a Red Hat rpm is
associated with, it needs to be able to find its repository in Red Hat's
published mapping file.

Even though some "arch-less" repository ids currently appear in the mapping file,
they are going to be removed soon and will be blocked by Konflux in the
future[2].

[0] https://security.access.redhat.com/data/meta/v1/repository-to-cpe.json
[1] https://cpe.mitre.org/
[2]

Signed-off-by: ralphbean <rbean@redhat.com>